requirements.txt
patchright==1.57.0
requests==2.31.0
pyyaml==6.0.1
colorama==0.4.6
python-dotenv==1.0.0

